+++ CHANGELOG
@@ -210,3 +210,4 @@ When SS inputs are selected in coincontrol:
 - allow to spend more than 0.1 as a fee (note: still will respect non-ss fee limits)
 Slight optimizations/readability improvements for DS 
 Unlock inputs on SS failure
+Move rand functions from util to new random.h/.cpp
