+++ src/key.cpp
@@ -12,6 +12,7 @@
 #include "crypto/common.h"
 #include "crypto/hmac_sha512.h"
 #include "pubkey.h"
+#include "random.h"
 #include "util.h"
 
 // anonymous namespace
