+++ src/makefile.bsd
@@ -120,6 +120,7 @@ OBJS= \
     obj/main.o \
     obj/net.o \
     obj/protocol.o \
+    obj/random.o \
     obj/rpc/rpcclient.o \
     obj/rpc/rpcprotocol.o \
     obj/rpc/rpcserver.o \
